bluetooth: host: audio: move audio assigned numbers to dedicated header
Move assigned number for audio from audio.h to assigned_numbers.h where
all assigned numbers are located.
